chore(minibend): rename source to datasource
